+++ xenon-common/src/test/java/com/vmware/xenon/services/common/MinimalTestService.java
@@ -129,9 +129,9 @@ public class MinimalTestService extends StatefulService {
         }
     }
 
-    public boolean gotStopped;
-
     public boolean gotDeleted;
+
+    public boolean gotStopped;
 
     @Override
     public void handleStop(Operation delete) {