+++ xenon-common/src/test/java/com/vmware/xenon/services/common/OnDemandLoadFactoryService.java
@@ -14,13 +14,43 @@
 package com.vmware.xenon.services.common;
 
 import java.util.EnumSet;
+import java.util.logging.Level;
 
 import com.vmware.xenon.common.FactoryService;
+import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.Service;
+import com.vmware.xenon.common.ServiceHost;
+import com.vmware.xenon.common.UriUtils;
+import com.vmware.xenon.common.test.TestContext;
 
 public class OnDemandLoadFactoryService extends FactoryService {
     public static final String SELF_LINK = "test/on-demand-load-services";
     public static final String FACTORY_LINK = "test/on-demand-load-services";
+
+    public static String create(ServiceHost h, ServiceOption... extraOptions)
+            throws Throwable {
+        // create an on demand load factory and services
+        TestContext ctx = TestContext.create(1, h.getOperationTimeoutMicros());
+        OnDemandLoadFactoryService s = new OnDemandLoadFactoryService();
+        EnumSet<ServiceOption> childOptions = EnumSet.of(ServiceOption.PERSISTENCE,
+                ServiceOption.REPLICATION, ServiceOption.OWNER_SELECTION);
+
+        if (extraOptions != null) {
+            for (ServiceOption eo : extraOptions) {
+                childOptions.add(eo);
+            }
+        }
+        s.setChildServiceCaps(childOptions);
+        Operation factoryPost = Operation.createPost(
+                UriUtils.buildUri(h, s.getClass()))
+                .setCompletion(ctx.getCompletion());
+        h.startService(factoryPost, s);
+        ctx.await();
+        String factoryLink = s.getSelfLink();
+        h.scheduleNodeGroupChangeMaintenance(ServiceUriPaths.DEFAULT_NODE_SELECTOR);
+        h.log(Level.INFO, "Started on demand load factory at %s", factoryLink);
+        return factoryLink;
+    }
 
     public OnDemandLoadFactoryService() {
         super(ExampleService.ExampleServiceState.class);